actualizacion_imagen

+++ database/migrations/2025_10_21_160855_add_code_goods_to_goods_table.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('goods', function (Blueprint $table) {
-            $table->dropColumn('code_goods');
+            if (Schema::hasColumn('goods', 'code_goods')) {
+                $table->dropColumn('code_goods');
+            }
         });
     }
 };
